+++ app/Http/Controllers/SkalaController.php
@@ -20,20 +20,22 @@ class SkalaController extends Controller
 
     public function accumulateMatrix()
     {
-        MatrixKeputusan::whereNotNull('id_matrix')->delete();
         // Matrix
         $kriteria = Kriteria::orderBy('id_kriteria', 'asc')->get();
         $alternatif = Alternatif::all();
         $bobot = Bobot::all();
         $skala = Skala::all();
 
-        for ($i = 0; $i < count($alternatif); $i++) {
-            for ($j = 0; $j < count($kriteria); $j++) {
-                MatrixKeputusan::create([
-                    'id_alternatif' => $alternatif[$i]->id_alternatif,
-                    'id_bobot' => $bobot[$j]->id_bobot,
-                    'id_skala' => $skala[$i]->id_skala,
-                ]);
+        if (count($alternatif) == count($skala)) {
+            MatrixKeputusan::whereNotNull('id_matrix')->delete();
+            for ($i = 0; $i < count($alternatif); $i++) {
+                for ($j = 0; $j < count($kriteria); $j++) {
+                    MatrixKeputusan::create([
+                        'id_alternatif' => $alternatif[$i]->id_alternatif,
+                        'id_bobot' => $bobot[$j]->id_bobot,
+                        'id_skala' => $skala[$i]->id_skala,
+                    ]);
+                }
             }
         }
     }
@@ -78,9 +80,9 @@ class SkalaController extends Controller
         try {
             DB::beginTransaction();
             $skala = Skala::find($id);
+            MatrixKeputusan::where('id_skala', $id)->delete();
             $skala->delete();
             DB::commit();
-            SkalaController::accumulateMatrix();
             return redirect()->back();
         } catch (\Exception $e) {
             DB::rollBack();
